+++ abdp.Service/Models/CommTrClassServiceModel.cs
@@ -8,10 +8,10 @@ namespace abdp.Service.Models
 {
     public class CommTrClassServiceModel
     {
-        public int group_code { get; set; }
+        public Int16 group_code { get; set; }
         public string class_code { get; set; }
         public string class_desc { get; set; }
-        public int? meter_size_code { get; set; }
+        public Int16? meter_size_code { get; set; }
         
         public string group_desc { get; set; }
     }
